Add new `customCss` dashboard option
